cloud_server/requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.21.0
pillow
numpy
curl-cffi
yfinance
mplfinance
pandas
python-multipart